// types.ts
export interface PurchaseLineItem {
    productId: string;
    productName: string;
    unit: string;
    quantity: number;
    price: number;
}

export interface Purchase {
    id: string;
    date: string;
    supplier: string;
    supplierContact: string;
    invoiceId: string;
    lineItems: PurchaseLineItem[];
    shippingCost?: number;
    otherFees?: number;
    total: number; // Grand total
    status: 'Paid' | 'Unpaid' | 'Overdue';
}
// ...
export interface ColumnDefinition<T> {
  header: string;
  accessor: keyof T;
  render?: (value: T[keyof T], row: T) => React.ReactNode;
  sortable?: boolean;
}